Implemented a MVC type architecture, authentication is now handled by one middleware and added an errror for when DB connection fails
- requireAuth is the only middleware required to check if the user is authorized and is logged in or not.
- Removed the ping and index paths
- All new routes from /api/auth/ redirected to /api/
- Fixed the gitignore file
