+++ .gitignore
@@ -19,5 +19,3 @@ yarn-debug.log*
 yarn-error.log*
 
 server/.env
-server/.env
-server/.env
