--- server/src/routes/index.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const router = express.Router();
-
-router.get('/welcome', (req, res) => {
-  res.status(200).send({ welcomeMessage: 'Step 1 (completed)' });
-});
-
-module.exports = router;
